--- bundles/representation-of-sales-permission/README.md
@@ -1,11 +1,0 @@
-# Representation Of Sales Permission
-[![CI](https://github.com/fond-of-oryx/representation-of-sales-permission/actions/workflows/main.yml/badge.svg)](https://github.com/fond-of-oryx/representation-of-sales-permission/actions/workflows/main.yml)
-[![license](https://img.shields.io/github/license/fond-of-oryx/representation-of-sales-permission.svg)](https://packagist.org/packages/fond-of-oryx/representation-of-sales-permission)
-
-## Installation
-
-```
-composer require fond-of-oryx/representation-of-sales-permission
-```
-
-Append permission in ZED Backend!
