--- bundles/representation-of-sales-permission/src/FondOfOryx/Client/RepresentationOfSalesPermission/Dependency/Client/RepresentationOfSalesPermissionToZedRequestBridge.php
@@ -1,34 +1,0 @@
-<?php
-
-namespace FondOfOryx\Client\RepresentationOfSalesPermission\Dependency\Client;
-
-use Spryker\Client\ZedRequest\ZedRequestClientInterface;
-use Spryker\Shared\Kernel\Transfer\TransferInterface;
-
-class RepresentationOfSalesPermissionToZedRequestBridge implements RepresentationOfSalesPermissionToZedRequestInterface
-{
-    /**
-     * @var \Spryker\Client\ZedRequest\ZedRequestClientInterface
-     */
-    protected $zedRequestClient;
-
-    /**
-     * @param \Spryker\Client\ZedRequest\ZedRequestClientInterface $zedRequestClient
-     */
-    public function __construct(ZedRequestClientInterface $zedRequestClient)
-    {
-        $this->zedRequestClient = $zedRequestClient;
-    }
-
-    /**
-     * @param string $url
-     * @param \Spryker\Shared\Kernel\Transfer\TransferInterface $object
-     * @param int|null $timeoutInSeconds
-     *
-     * @return \Spryker\Shared\Kernel\Transfer\TransferInterface
-     */
-    public function call(string $url, TransferInterface $object, ?int $timeoutInSeconds = null): TransferInterface
-    {
-        return $this->zedRequestClient->call($url, $object, $timeoutInSeconds);
-    }
-}
