--- bundles/representation-of-sales-permission/src/FondOfOryx/Shared/RepresentationOfSalesPermission/Transfer/representation_of_sales_permission.transfer.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-<transfers xmlns="spryker:transfer-01"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="spryker:transfer-01 http://static.spryker.com/transfer-01.xsd">
-
-    <transfer name="RepresentationOfSalesPermissionResponse">
-        <property name="request" type="RepresentationOfSalesPermissionRequest" />
-        <property name="hasPermissionToManageOwnRepresentations" type="bool" />
-        <property name="hasPermissionToManageGlobalRepresentations" type="bool" />
-    </transfer>
-
-    <transfer name="RepresentationOfSalesPermissionRequest">
-        <property name="permissionKey" type="string" />
-        <property name="originatorReference" type="string" />
-        <property name="distributorReference" type="string" />
-    </transfer>
-</transfers>
