--- bundles/representation-of-sales-permission/src/FondOfOryx/Zed/RepresentationOfSalesPermission/Communication/Plugin/Permission/CanManageRepresentationGlobalPermissionPlugin.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types = 1);
-
-namespace FondOfOryx\Zed\RepresentationOfSalesPermission\Communication\Plugin\Permission;
-
-use Spryker\Shared\PermissionExtension\Dependency\Plugin\PermissionPluginInterface;
-
-class CanManageRepresentationGlobalPermissionPlugin implements PermissionPluginInterface
-{
-    /**
-     * @var string
-     */
-    public const KEY = 'CanManageRepresentationGlobalPermissionPlugin';
-
-    /**
-     * @return string
-     */
-    public function getKey(): string
-    {
-        return static::KEY;
-    }
-}
