+++ bundles/representative-company-user-rest-api-permission/README.md
@@ -1,0 +1,11 @@
+# Representation Of Sales Permission
+[![CI](https://github.com/fond-of-oryx/representative-company-user-rest-api-permission/actions/workflows/main.yml/badge.svg)](https://github.com/fond-of-oryx/representative-company-user-rest-api-permission/actions/workflows/main.yml)
+[![license](https://img.shields.io/github/license/fond-of-oryx/representative-company-user-rest-api-permission.svg)](https://packagist.org/packages/fond-of-oryx/representative-company-user-rest-api-permission)
+
+## Installation
+
+```
+composer require fond-of-oryx/representative-company-user-rest-api-permission
+```
+
+Append permission in ZED Backend!
