+++ bundles/representative-company-user-rest-api-permission/composer.json
@@ -1,0 +1,42 @@
+{
+  "name": "fond-of-oryx/representative-company-user-rest-api-permission",
+  "description": "Representative Company User Rest API module",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Julian Hyatt",
+      "email": "info@jzit.de"
+    }
+  ],
+  "require": {
+    "php": ">=8.0",
+    "spryker/company-role": "^1.0.0",
+    "spryker/permission-extension": "^1.0.0",
+    "spryker/uuid-behavior": "^1.0.0"
+  },
+  "require-dev": {
+    "fond-of-codeception/spryker": "*",
+    "spryker/code-sniffer": "*",
+    "spryker-sdk/phpstan-spryker": "*"
+  },
+  "minimum-stability": "dev",
+  "prefer-stable": true,
+  "autoload": {
+    "psr-4": {
+      "FondOfOryx\\": "src/FondOfOryx/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "FondOfOryx\\": "tests/FondOfOryx/",
+      "Generated\\": "src/Generated/",
+      "Orm\\Zed\\": "src/Orm/Zed/"
+    }
+  },
+  "config": {
+    "sort-packages": true,
+    "allow-plugins": {
+      "dealerdirect/phpcodesniffer-composer-installer": true
+    }
+  }
+}
