+++ bundles/representative-company-user-rest-api-permission/phpstan.neon
