+++ bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Client/RepresentativeCompanyUserRestApiPermission/Dependency/Client/RepresentativeCompanyUserRestApiPermissionToZedRequestBridge.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace FondOfOryx\Client\RepresentativeCompanyUserRestApiPermission\Dependency\Client;
+
+use Spryker\Client\ZedRequest\ZedRequestClientInterface;
+use Spryker\Shared\Kernel\Transfer\TransferInterface;
+
+class RepresentativeCompanyUserRestApiPermissionToZedRequestBridge implements RepresentativeCompanyUserRestApiPermissionToZedRequestInterface
+{
+    /**
+     * @var \Spryker\Client\ZedRequest\ZedRequestClientInterface
+     */
+    protected $zedRequestClient;
+
+    /**
+     * @param \Spryker\Client\ZedRequest\ZedRequestClientInterface $zedRequestClient
+     */
+    public function __construct(ZedRequestClientInterface $zedRequestClient)
+    {
+        $this->zedRequestClient = $zedRequestClient;
+    }
+
+    /**
+     * @param string $url
+     * @param \Spryker\Shared\Kernel\Transfer\TransferInterface $object
+     * @param int|null $timeoutInSeconds
+     *
+     * @return \Spryker\Shared\Kernel\Transfer\TransferInterface
+     */
+    public function call(string $url, TransferInterface $object, ?int $timeoutInSeconds = null): TransferInterface
+    {
+        return $this->zedRequestClient->call($url, $object, $timeoutInSeconds);
+    }
+}
