+++ bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Client/RepresentativeCompanyUserRestApiPermission/RepresentativeCompanyUserRestApiPermissionClientInterface.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace FondOfOryx\Client\RepresentativeCompanyUserRestApiPermission;
+
+use Generated\Shared\Transfer\RepresentativeCompanyUserRestApiPermissionRequestTransfer;
+
+interface RepresentativeCompanyUserRestApiPermissionClientInterface
+{
+    /**
+     * @param \Generated\Shared\Transfer\RepresentativeCompanyUserRestApiPermissionRequestTransfer $representationOfSalesPermissionRequestTransfer
+     *
+     * @throws \Exception
+     *
+     * @return bool
+     */
+    public function hasPermissionToManageOwnRepresentations(
+        RepresentativeCompanyUserRestApiPermissionRequestTransfer $representationOfSalesPermissionRequestTransfer
+    ): bool;
+
+    /**
+     * @param \Generated\Shared\Transfer\RepresentativeCompanyUserRestApiPermissionRequestTransfer $representationOfSalesPermissionRequestTransfer
+     *
+     * @throws \Exception
+     *
+     * @return bool
+     */
+    public function hasPermissionToManageGlobalRepresentations(
+        RepresentativeCompanyUserRestApiPermissionRequestTransfer $representationOfSalesPermissionRequestTransfer
+    ): bool;
+}
