+++ bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Client/RepresentativeCompanyUserRestApiPermission/RepresentativeCompanyUserRestApiPermissionDependencyProvider.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace FondOfOryx\Client\RepresentativeCompanyUserRestApiPermission;
+
+use FondOfOryx\Client\RepresentativeCompanyUserRestApiPermission\Dependency\Client\RepresentativeCompanyUserRestApiPermissionToZedRequestBridge;
+use Spryker\Client\Kernel\AbstractDependencyProvider;
+use Spryker\Client\Kernel\Container;
+
+class RepresentativeCompanyUserRestApiPermissionDependencyProvider extends AbstractDependencyProvider
+{
+    /**
+     * @var string
+     */
+    public const CLIENT_ZED_REQUEST = 'CLIENT_ZED_REQUEST';
+
+    /**
+     * @param \Spryker\Client\Kernel\Container $container
+     *
+     * @return \Spryker\Client\Kernel\Container
+     */
+    public function provideServiceLayerDependencies(Container $container)
+    {
+        $container = $this->addZedRequestClient($container);
+
+        return $container;
+    }
+
+    /**
+     * @param \Spryker\Client\Kernel\Container $container
+     *
+     * @return \Spryker\Client\Kernel\Container
+     */
+    protected function addZedRequestClient(Container $container)
+    {
+        $container[static::CLIENT_ZED_REQUEST] = function (Container $container) {
+            return new RepresentativeCompanyUserRestApiPermissionToZedRequestBridge($container->getLocator()->zedRequest()->client());
+        };
+
+        return $container;
+    }
+}
