+++ bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Client/RepresentativeCompanyUserRestApiPermission/Zed/RepresentativeCompanyUserRestApiPermissionStubInterface.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace FondOfOryx\Client\RepresentativeCompanyUserRestApiPermission\Zed;
+
+use Generated\Shared\Transfer\RepresentativeCompanyUserRestApiPermissionRequestTransfer;
+
+interface RepresentativeCompanyUserRestApiPermissionStubInterface
+{
+    /**
+     * @param \Generated\Shared\Transfer\RepresentativeCompanyUserRestApiPermissionRequestTransfer $representationOfSalesPermissionRequestTransfer
+     *
+     * @return \Spryker\Shared\Kernel\Transfer\TransferInterface
+     */
+    public function hasPermissionToManageOwnRepresentations(
+        RepresentativeCompanyUserRestApiPermissionRequestTransfer $representationOfSalesPermissionRequestTransfer
+    );
+
+    /**
+     * @param \Generated\Shared\Transfer\RepresentativeCompanyUserRestApiPermissionRequestTransfer $representationOfSalesPermissionRequestTransfer
+     *
+     * @return \Spryker\Shared\Kernel\Transfer\TransferInterface
+     */
+    public function hasPermissionToManageGlobalRepresentations(
+        RepresentativeCompanyUserRestApiPermissionRequestTransfer $representationOfSalesPermissionRequestTransfer
+    );
+}
