+++ bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Shared/RepresentativeCompanyUserRestApiPermission/Transfer/representative_company_user_rest_api_permission.transfer.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0"?>
+<transfers xmlns="spryker:transfer-01"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="spryker:transfer-01 http://static.spryker.com/transfer-01.xsd">
+
+    <transfer name="RepresentativeCompanyUserRestApiPermissionResponse">
+        <property name="request" type="RepresentativeCompanyUserRestApiPermissionRequest" />
+        <property name="hasPermissionToManageOwnRepresentations" type="bool" />
+        <property name="hasPermissionToManageGlobalRepresentations" type="bool" />
+    </transfer>
+
+    <transfer name="RepresentativeCompanyUserRestApiPermissionRequest">
+        <property name="permissionKey" type="string" />
+        <property name="originatorReference" type="string" />
+        <property name="distributorReference" type="string" />
+    </transfer>
+</transfers>
