+++ bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Zed/RepresentativeCompanyUserRestApiPermission/Communication/Plugin/Permission/CanManageRepresentationPermissionPlugin.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace FondOfOryx\Zed\RepresentativeCompanyUserRestApiPermission\Communication\Plugin\Permission;
+
+use Spryker\Shared\PermissionExtension\Dependency\Plugin\PermissionPluginInterface;
+
+class CanManageRepresentationPermissionPlugin implements PermissionPluginInterface
+{
+    /**
+     * @var string
+     */
+    public const KEY = 'CanManageRepresentationPermissionPlugin';
+
+    /**
+     * @return string
+     */
+    public function getKey(): string
+    {
+        return static::KEY;
+    }
+}
