+++ bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Zed/RepresentativeCompanyUserRestApiPermission/Persistence/RepresentativeCompanyUserRestApiPermissionPersistenceFactory.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace FondOfOryx\Zed\RepresentativeCompanyUserRestApiPermission\Persistence;
+
+use FondOfOryx\Zed\RepresentativeCompanyUserRestApiPermission\RepresentativeCompanyUserRestApiPermissionDependencyProvider;
+use Orm\Zed\CompanyRole\Persistence\SpyCompanyRoleToPermissionQuery;
+use Spryker\Zed\Kernel\Persistence\AbstractPersistenceFactory;
+
+class RepresentativeCompanyUserRestApiPermissionPersistenceFactory extends AbstractPersistenceFactory
+{
+    /**
+     * @return \Orm\Zed\CompanyRole\Persistence\SpyCompanyRoleToPermissionQuery
+     */
+    public function getSpyCompanyRoleToPermissionQuery(): SpyCompanyRoleToPermissionQuery
+    {
+        return $this->getProvidedDependency(RepresentativeCompanyUserRestApiPermissionDependencyProvider::QUERY_SPY_COMPANY_ROLE_PERMISSION);
+    }
+}
