+++ bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Zed/RepresentativeCompanyUserRestApiPermission/Persistence/RepresentativeCompanyUserRestApiPermissionRepositoryInterface.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace FondOfOryx\Zed\RepresentativeCompanyUserRestApiPermission\Persistence;
+
+interface RepresentativeCompanyUserRestApiPermissionRepositoryInterface
+{
+    /**
+     * @param string $permissionKey
+     * @param string $customerReference
+     *
+     * @throws \Spryker\Zed\Propel\Business\Exception\AmbiguousComparisonException
+     *
+     * @return bool
+     */
+    public function hasPermission(
+        string $permissionKey,
+        string $customerReference
+    ): bool;
+}
