+++ composer.json
@@ -314,7 +314,7 @@
     "fond-of-oryx/product-payment-restriction": "*",
     "fond-of-oryx/qr-code-generator": "*",
     "fond-of-oryx/representation-of-sales-extension": "*",
-    "fond-of-oryx/representation-of-sales-permission": "*",
+    "fond-of-oryx/representative-company-user-rest-api-permission": "*",
     "fond-of-oryx/return-label": "*",
     "fond-of-oryx/return-labels-rest-api": "*",
     "fond-of-oryx/return-labels-rest-api-company-connector": "*",
@@ -392,7 +392,7 @@
       "FondOfOryx\\Client\\ProductLocaleRestrictionStorage\\": "bundles/product-locale-restriction-storage/src/FondOfOryx/Client/ProductLocaleRestrictionStorage",
       "FondOfOryx\\Client\\ProductListsRestApi\\": "bundles/product-lists-rest-api/src/FondOfOryx/Client/ProductListsRestApi",
       "FondOfOryx\\Client\\ProductListSearchRestApi\\": "bundles/product-list-search-rest-api/src/FondOfOryx/Client/ProductListSearchRestApi",
-      "FondOfOryx\\Client\\RepresentationOfSalesPermission\\": "bundles/representation-of-sales-permission/src/FondOfOryx/Client/RepresentationOfSalesPermission",
+      "FondOfOryx\\Client\\RepresentativeCompanyUserRestApiPermission\\": "bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Client/RepresentativeCompanyUserRestApiPermission",
       "FondOfOryx\\Client\\ReturnLabelsRestApi\\": "bundles/return-labels-rest-api/src/FondOfOryx/Client/ReturnLabelsRestApi",
       "FondOfOryx\\Client\\CustomerTokenManager\\": "bundles/customer-token-manager/src/FondOfOryx/Client/CustomerTokenManager",
       "FondOfOryx\\Client\\SplittableCheckoutRestApi\\": "bundles/splittable-checkout-rest-api/src/FondOfOryx/Client/SplittableCheckoutRestApi",
@@ -534,7 +534,7 @@
       "FondOfOryx\\Shared\\ProductListSearchRestApiExtension\\": "bundles/product-list-search-rest-api-extension/src/FondOfOryx/Shared/ProductListSearchRestApiExtension",
       "FondOfOryx\\Shared\\ProductLocaleRestrictionStorage\\": "bundles/product-locale-restriction-storage/src/FondOfOryx/Shared/ProductLocaleRestrictionStorage",
       "FondOfOryx\\Shared\\QrCodeGenerator\\": "bundles/qr-code-generator/src/FondOfOryx/Shared/QrCodeGenerator",
-      "FondOfOryx\\Shared\\RepresentationOfSalesPermission\\": "bundles/representation-of-sales-permission/src/FondOfOryx/Shared/RepresentationOfSalesPermission",
+      "FondOfOryx\\Shared\\RepresentativeCompanyUserRestApiPermission\\": "bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Shared/RepresentativeCompanyUserRestApiPermission",
       "FondOfOryx\\Shared\\ReturnLabel\\": "bundles/return-label/src/FondOfOryx/Shared/ReturnLabel",
       "FondOfOryx\\Shared\\ReturnLabelsRestApi\\": "bundles/return-labels-rest-api/src/FondOfOryx/Shared/ReturnLabelsRestApi",
       "FondOfOryx\\Shared\\ReturnLabelsRestApiCompanyConnector\\": "bundles/return-labels-rest-api-company-connector/src/FondOfOryx/Shared/ReturnLabelsRestApiCompanyConnector",
@@ -738,7 +738,7 @@
       "FondOfOryx\\Zed\\ProductLocaleRestrictionStorage\\": "bundles/product-locale-restriction-storage/src/FondOfOryx/Zed/ProductLocaleRestrictionStorage",
       "FondOfOryx\\Zed\\ProductLocaleRestriction\\": "bundles/product-locale-restriction/src/FondOfOryx/Zed/ProductLocaleRestriction",
       "FondOfOryx\\Zed\\RepresentationOfSalesExtension\\": "bundles/representation-of-sales-extension/src/FondOfOryx/Zed/RepresentationOfSalesExtension",
-      "FondOfOryx\\Zed\\RepresentationOfSalesPermission\\": "bundles/representation-of-sales-permission/src/FondOfOryx/Zed/RepresentationOfSalesPermission",
+      "FondOfOryx\\Zed\\RepresentativeCompanyUserRestApiPermission\\": "bundles/representative-company-user-rest-api-permission/src/FondOfOryx/Zed/RepresentativeCompanyUserRestApiPermission",
       "FondOfOryx\\Zed\\ReturnLabel\\": "bundles/return-label/src/FondOfOryx/Zed/ReturnLabel",
       "FondOfOryx\\Zed\\ReturnLabelsRestApi\\": "bundles/return-labels-rest-api/src/FondOfOryx/Zed/ReturnLabelsRestApi",
       "FondOfOryx\\Zed\\ReturnLabelsRestApiCompanyBusinessUnitConnector\\": "bundles/return-labels-rest-api-company-business-unit-connector/src/FondOfOryx/Zed/ReturnLabelsRestApiCompanyBusinessUnitConnector",
