+++ dandelion.json
@@ -747,8 +747,8 @@
       "path": "bundles/qr-code-generator/",
       "version": "2.0.0"
     },
-    "representation-of-sales-permission": {
-      "path": "bundles/representation-of-sales-permission/",
+    "representative-company-user-rest-api-permission": {
+      "path": "bundles/representative-company-user-rest-api-permission/",
       "version": "1.0.0.beta.1"
     },
     "return-label": {
